feat: complete the course with exe 15 to 18
