--- src/16-omit-and-pick.solution.2.ts
@@ -1,16 +1,0 @@
-import { Equal, Expect } from "./helpers/type-utils";
-
-interface User {
-  id: string;
-  firstName: string;
-  lastName: string;
-}
-
-/**
- * How do we create a new object type with _only_ the
- * firstName and lastName properties of User?
- */
-
-type MyType = Pick<User, "firstName" | "lastName">;
-
-type tests = [Expect<Equal<MyType, { firstName: string; lastName: string }>>];
