+++ src/done/15-intersection.problem.ts
@@ -14,7 +14,11 @@ interface Post {
  * How do we type this return statement so it's both
  * User AND { posts: Post[] }
  */
-export const getDefaultUserAndPosts = (): unknown => {
+
+type UserAndPosts = User & { posts: Post[] };
+
+// export const getDefaultUserAndPosts = (): User & { posts: Post[] } => {
+export const getDefaultUserAndPosts = (): UserAndPosts => {
   return {
     id: "1",
     firstName: "Matt",
