+++ src/done/15-intersection.solution.ts
