+++ src/done/16-omit-and-pick.solution.1.ts
@@ -1,4 +1,4 @@
-import { Equal, Expect } from "./helpers/type-utils";
+import { Equal, Expect } from "../helpers/type-utils";
 
 interface User {
   id: string;
