+++ src/done/17-function-types.problem.ts
@@ -1,9 +1,9 @@
-import { Equal, Expect } from "./helpers/type-utils";
+import { Equal, Expect } from "../helpers/type-utils";
 
 /**
  * How do we type onFocusChange?
  */
-const addListener = (onFocusChange: (isFocused: boolean) => void) => {
+const addListener = (onFocusChange: (x: boolean) => void) => {
   window.addEventListener("focus", () => {
     onFocusChange(true);
   });
