+++ src/done/17-function-types.solution.1.ts
@@ -1,11 +1,9 @@
-import { Equal, Expect } from "./helpers/type-utils";
+import { Equal, Expect } from "../helpers/type-utils";
 
 /**
  * How do we type onFocusChange?
  */
-type FocusListener = (isFocused: boolean) => void;
-
-const addListener = (onFocusChange: FocusListener) => {
+const addListener = (onFocusChange: (isFocused: boolean) => void) => {
   window.addEventListener("focus", () => {
     onFocusChange(true);
   });
