+++ src/done/18-function-types-with-promises.solution.ts
